refactor(models): move the pricing tier editors under PricingSection (LFE-14948)
'pricing-tiers' was not a kind folder (structure RFC rule 5); the editors are
PricingSection's private subcomponents, so they live in its component folder.
